# api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /event/events/search:
    put:
      tags:
      - event-controller
      operationId: filterEvents
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Filter"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /event/event/phil_flag_id={id}_num=_{num}:
    put:
      tags:
      - event-controller
      operationId: philFlagEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: num
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Event"
  /event/event/fee_flag_id={id}_num={num}:
    put:
      tags:
      - event-controller
      operationId: feeFlagEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: num
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Event"
  /event/event/edit_{id}:
    put:
      tags:
      - event-controller
      operationId: editEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Event"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Event"
  /event/event/delete_{id}:
    put:
      tags:
      - event-controller
      operationId: deleteEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Event"
  /event/event/conf_flag_id={id}_num={num}:
    put:
      tags:
      - event-controller
      operationId: confFlagEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: num
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Event"
...
  /admin/groups/finance_groups/search:
    put:
      tags:
      - finance-group-controller
      operationId: filterFinanceGroups
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Filter"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/groups/finance_group_{id}:
    put:
      tags:
      - finance-group-controller
      operationId: removeGroup
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FinanceGroup"
...
  /event/event:
    post:
      tags:
      - event-controller
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Event"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Event"
...
  /admin/groups/finance_group:
    post:
      tags:
      - finance-group-controller
      operationId: createGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceGroup"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FinanceGroup"
...
components:
  schemas:
    Filter:
      type: object
      properties:
        col:
          type: string
        op:
          type: string
        val:
          type: object
...
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        deleted:
          type: integer
          format: int32
        feeFlag:
          type: integer
          format: int32
        estAttendance:
          type: integer
          format: int32
        philanthropyFlag:
          type: integer
          format: int32
        conferenceFlag:
          type: integer
          format: int32
...
    FinanceGroup:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        deleted:
          type: integer
          format: int32